thymatic-health/backend/pyproject.toml
[project]
name = "thymatic-health-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "thymia-sentinel>=0.1",
    "google-genai>=1.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "websockets>=12",
    "numpy>=1.26",
    "mediapipe>=0.10",
    "opencv-python-headless>=4.9",
    "fal-client>=0.4",
    "python-multipart>=0.0.9",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
